hashing the secret this time
