--- src/pages/api/v1/pull-requests/index.ts
@@ -1,3 +1,0 @@
-import handlePullRequest from "./handlePullRequest";
-
-export default handlePullRequest;
